Training bot on Uni Data
Training the chatterbot on Uni data can be done in conversational style. This is an enormous task however, as the volume of examples needs to be very high.

Thus a new solution using openAI's model has been proposed in addition, being able to be trained on much shorter, descriptive texts.